Removed some components of the classes and moved it back to main file

+++ TrebleSketch-AIE-Asteroids/ShipClass.cs
@@ -17,7 +17,7 @@ namespace TrebleSketch_AIE_Asteroids
         public float RotationDelta;
 
         public Vector2 Size;
-        public int Radius;
+        public float Radius;
         public Vector2 MaxLimit;
         public Vector2 MinLimit;
